Upgrade syntax to sql`query here`
This leads to less verbosity in the need for delimiters: sql`x` is
shorter than @query("x"). Crucially, editor syntax highlighers seem to
highlight all $ interpolations within Cmd-style backticks, so we still
get syntax highlighting with this in contrast to a normal string macro.

Now that we have more control over parsing, upgrade the parsing rules so
that SQL single-quoted strings are handled specifically and allow $ to
be written without escapes.